requirements.txt
opencv-python
pillow
pyperclip
zxing-cpp